Changes to how the heapstack works

+++ src/platform.h
@@ -1,6 +1,5 @@
 #pragma once
 
-#include <cstddef>
 float platform_memory_mb();
 
 void* platform_alloc_memory(unsigned long long amount);
@@ -9,7 +8,7 @@ void platform_free_memory(void* pointer, unsigned long long size);
 
 void* load_entire_file(const char* filepath);
 
-size_t get_file_size(const char* filepath);
+unsigned long long get_file_size(const char* filepath);
 
 //void free_file(const char* filepath);
